feat(ui): 優化導航與用戶體驗功能
- 建立共享 AppNavbar 組件，統一所有頁面導航體驗
- 個人資料頁面新增拖曳上傳頭像功能，支援即時預覽與檔案驗證
- 成功訊息改為右下角 Toast 樣式顯示，提升用戶體驗
- 修復 AdminController 認證問題，確保管理員 API 正常運作
- 所有頁面維持一致的下拉選單導航，移除返回按鈕設計
