+++ resources/js/components/AppNavbar.vue
@@ -1,0 +1,235 @@
+<template>
+  <nav class="bg-white shadow">
+    <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+      <div class="flex justify-between h-16">
+        <div class="flex items-center">
+          <router-link to="/dashboard" class="flex-shrink-0 flex items-center">
+            <div class="h-8 w-8 bg-primary-500 rounded-full flex items-center justify-center">
+              <svg class="h-5 w-5 text-white" fill="currentColor" viewBox="0 0 20 20">
+                <path fill-rule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.857-9.809a.75.75 0 00-1.214-.882l-3.236 4.53L8.93 10.7a.75.75 0 00-1.06 1.061l1.5 1.5a.75.75 0 001.137-.089l4-5.5z" clip-rule="evenodd" />
+              </svg>
+            </div>
+            <span class="ml-2 text-xl font-bold text-gray-900">PurpleDesk</span>
+            <span v-if="isAdmin" class="ml-2 px-2 py-1 text-xs bg-red-100 text-red-800 rounded-full">管理員</span>
+          </router-link>
+        </div>
+        
+        <!-- 個人選單 -->
+        <div class="flex items-center space-x-4">
+          <!-- 通知圖示 -->
+          <button class="text-gray-500 hover:text-gray-700 p-1 rounded-full hover:bg-gray-100">
+            <svg class="h-5 w-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 17h5l-3.405-3.405A2.032 2.032 0 0116 14.158V11a6.002 6.002 0 00-4-5.659V5a2 2 0 10-4 0v.341C6.67 6.165 6 8.388 6 11v3.159c0 .538-.214 1.055-.595 1.436L2 19h5m6 0v1a3 3 0 11-6 0v-1m6 0H9"></path>
+            </svg>
+          </button>
+
+          <!-- 個人選單下拉 -->
+          <div class="relative">
+            <button
+              @click="toggleUserMenu"
+              class="flex items-center space-x-2 p-1 rounded-full hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-primary-500 focus:ring-offset-2"
+            >
+              <!-- 使用者頭像 -->
+              <div class="h-8 w-8 rounded-full bg-primary-500 flex items-center justify-center overflow-hidden">
+                <img
+                  v-if="user?.avatar"
+                  :src="user.avatar"
+                  :alt="user.name"
+                  class="h-full w-full object-cover"
+                />
+                <span v-else class="text-white text-sm font-medium">
+                  {{ getUserInitials(user) }}
+                </span>
+              </div>
+              
+              <!-- 下拉箭頭 -->
+              <svg class="h-4 w-4 text-gray-500" :class="{ 'rotate-180': showUserMenu }" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path>
+              </svg>
+            </button>
+
+            <!-- 下拉選單 -->
+            <div
+              v-show="showUserMenu"
+              class="absolute right-0 mt-2 w-64 bg-white rounded-md shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none z-50"
+              @click.away="showUserMenu = false"
+            >
+              <!-- 使用者資訊 -->
+              <div class="px-4 py-3 border-b border-gray-200">
+                <div class="flex items-center space-x-3">
+                  <div class="h-10 w-10 rounded-full bg-primary-500 flex items-center justify-center overflow-hidden">
+                    <img
+                      v-if="user?.avatar"
+                      :src="user.avatar"
+                      :alt="user.name"
+                      class="h-full w-full object-cover"
+                    />
+                    <span v-else class="text-white font-medium">
+                      {{ getUserInitials(user) }}
+                    </span>
+                  </div>
+                  <div class="flex-1 min-w-0">
+                    <p class="text-sm font-medium text-gray-900 truncate">
+                      {{ user?.display_name || user?.name }}
+                    </p>
+                    <p class="text-sm text-gray-500 truncate">
+                      {{ user?.email }}
+                    </p>
+                    <p v-if="user?.organization" class="text-xs text-gray-400 truncate">
+                      {{ user.organization.name }}
+                    </p>
+                  </div>
+                </div>
+              </div>
+
+              <!-- 選單項目 -->
+              <div class="py-1">
+                <!-- 個人資料 -->
+                <router-link
+                  to="/profile"
+                  class="flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
+                  @click="showUserMenu = false"
+                >
+                  <svg class="mr-3 h-4 w-4 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z"></path>
+                  </svg>
+                  個人資料
+                </router-link>
+
+                <!-- 設定 -->
+                <router-link
+                  to="/settings"
+                  class="flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
+                  @click="showUserMenu = false"
+                >
+                  <svg class="mr-3 h-4 w-4 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z"></path>
+                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"></path>
+                  </svg>
+                  設定
+                </router-link>
+
+                <!-- 管理員專用選項 -->
+                <template v-if="isAdmin">
+                  <div class="border-t border-gray-200 my-1"></div>
+                  
+                  <router-link
+                    to="/admin/users"
+                    class="flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
+                    @click="showUserMenu = false"
+                  >
+                    <svg class="mr-3 h-4 w-4 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 4.354a4 4 0 110 5.292M15 21H3v-1a6 6 0 0112 0v1zm0 0h6v-1a6 6 0 00-9-5.197m13.5-9a2.5 2.5 0 11-5 0 2.5 2.5 0 015 0z"></path>
+                    </svg>
+                    使用者管理
+                  </router-link>
+                  
+                  <router-link
+                    to="/admin/organizations"
+                    class="flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
+                    @click="showUserMenu = false"
+                  >
+                    <svg class="mr-3 h-4 w-4 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 21V5a2 2 0 00-2-2H7a2 2 0 00-2 2v16m14 0h2m-2 0h-5m-9 0H3m2 0h5M9 7h1m-1 4h1m4-4h1m-1 4h1m-5 10v-5a1 1 0 011-1h2a1 1 0 011 1v5m-4 0h4"></path>
+                    </svg>
+                    單位管理
+                  </router-link>
+                  
+                  <router-link
+                    to="/admin/system"
+                    class="flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
+                    @click="showUserMenu = false"
+                  >
+                    <svg class="mr-3 h-4 w-4 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                      <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 3v2m6-2v2M9 19v2m6-2v2M5 9H3m2 6H3m18-6h-2m2 6h-2M7 19h10a2 2 0 002-2V7a2 2 0 00-2-2H7a2 2 0 00-2 2v10a2 2 0 002 2zM9 9h6v6H9V9z"></path>
+                    </svg>
+                    系統設定
+                  </router-link>
+                </template>
+
+                <div class="border-t border-gray-200 my-1"></div>
+                
+                <!-- 登出 -->
+                <button
+                  @click="handleLogout"
+                  class="flex items-center w-full px-4 py-2 text-sm text-gray-700 hover:bg-gray-100 text-left"
+                >
+                  <svg class="mr-3 h-4 w-4 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M17 16l4-4m0 0l-4-4m4 4H7m6 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h4a3 3 0 013 3v1"></path>
+                  </svg>
+                  登出
+                </button>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </nav>
+</template>
+
+<script>
+import { computed, ref, onMounted, onUnmounted } from 'vue'
+import { useRouter } from 'vue-router'
+import { useAuthStore } from '../stores/auth'
+
+export default {
+  name: 'AppNavbar',
+  setup() {
+    const router = useRouter()
+    const authStore = useAuthStore()
+    const showUserMenu = ref(false)
+    
+    const user = computed(() => authStore.user)
+    
+    // 判斷是否為管理員
+    const isAdmin = computed(() => {
+      return user.value?.email === 'admin@purpledesk.com'
+    })
+    
+    const getUserInitials = (user) => {
+      if (!user) return ''
+      const name = user.display_name || user.name
+      return name.split(' ').map(n => n[0]).join('').toUpperCase().slice(0, 2)
+    }
+    
+    const toggleUserMenu = () => {
+      showUserMenu.value = !showUserMenu.value
+    }
+    
+    const handleLogout = async () => {
+      try {
+        showUserMenu.value = false
+        await authStore.logout()
+        router.push('/login')
+      } catch (error) {
+        console.error('登出錯誤:', error)
+      }
+    }
+    
+    // 點擊外部關閉選單
+    const handleClickOutside = (event) => {
+      if (!event.target.closest('.relative')) {
+        showUserMenu.value = false
+      }
+    }
+    
+    onMounted(() => {
+      document.addEventListener('click', handleClickOutside)
+    })
+    
+    onUnmounted(() => {
+      document.removeEventListener('click', handleClickOutside)
+    })
+    
+    return {
+      user,
+      isAdmin,
+      showUserMenu,
+      getUserInitials,
+      toggleUserMenu,
+      handleLogout
+    }
+  }
+}
+</script>
